refact(auth): remove auth from user module
