+++ src/modules/auth/schemas/models/auth.interface.ts
@@ -1,0 +1,4 @@
+export interface AuthCredentials {
+  email: string
+  password: string
+}
