+++ src/modules/auth/schemas/zod-validation/login.zod-validation.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod'
+
+export const loginSchema = z.object({
+  email: z.string(),
+  password: z.string()
+})
