+++ src/modules/user/schemas/models/user.interface.ts
@@ -9,10 +9,5 @@ export interface InterfaceUser {
   email: string
 }
 
-export interface UserCredentials {
-  email: string
-  password: string
-}
-
 // returned on public routes
 export type PublicInterfaceUser = Omit<InterfaceUser, 'password'>
